getMateriasAprobadas(), obtenerMateriasAprobadasParaUnAlumno(), obtenerMateriasQueFaltanCursarParaUnAlumno(), calcularPromedio()

+++ src/ar/edu/unlam/pb2/Parcial01/Evaluacion.java
@@ -3,12 +3,12 @@ package ar.edu.unlam.pb2.Parcial01;
 public class Evaluacion {
 	
 	private TipoDeNota tipoDeNota;
-	private Integer nota;
+	private Double nota;
 	private EstadoExamen estado;
 	private Materia materia;
 	private Alumno alumno;
 
-	public Evaluacion (TipoDeNota tipoDeNota, Integer nota, Materia materia, Alumno alumno) {
+	public Evaluacion (TipoDeNota tipoDeNota, Double nota, Materia materia, Alumno alumno) {
 		this.tipoDeNota=tipoDeNota;
 		this.nota=nota;
 		this.estado=this.estadoDelExamen(nota);
@@ -24,11 +24,11 @@ public class Evaluacion {
 		this.tipoDeNota = tipoDeNota;
 	}
 
-	public Integer getNota() {
+	public Double getNota() {
 		return nota;
 	}
 
-	public void setNota(Integer nota) {
+	public void setNota(Double nota) {
 		this.nota = nota;
 	}
 	
@@ -60,11 +60,11 @@ public class Evaluacion {
 		this.alumno = alumno;
 	}
 
-	public EstadoExamen estadoDelExamen(Integer nota) {
+	public EstadoExamen estadoDelExamen(Double nota) {
 		EstadoExamen estadoDelExamen = null;
-		if(nota < 4) {
+		if(nota < 4.0) {
 			estadoDelExamen = estadoDelExamen.desaprobado;
-		} else if(nota >= 4 && nota<7) {
+		} else if(nota >= 4.0 && nota<7.0) {
 			estadoDelExamen = estadoDelExamen.aprobado;
 		} else {
 			estadoDelExamen = estadoDelExamen.promocion;
